(feat) Transferred app component to App.jsx
