+++ client/app/app.jsx
@@ -1,21 +1,30 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-// Get main React component.
-var App = require('./index.jsx');
 
-//If the user allowed the site to access geolocation,
-//then find the user's location and pass it to the app.
-if (navigator.geolocation) {
-  navigator.geolocation.getCurrentPosition((position) => {
-  	var pos = position.coords.latitude + ', ' + position.coords.longitude;
-    ReactDOM.render(<App startLoc={pos}/>, document.getElementById('app'));
-  });
-//If the user did not allow the page to access geolocation,
-//then default to San Francisco.
-} else {
-	console.log('Geolocation not enabled by browser');
-	ReactDOM.render(<App startLoc="SF"/>, document.getElementById('app'));
-}
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      location: 'San Francisco'
+    };
+  }
 
+  setLocation(loc) {
+    this.setState({
+      location: loc
+    });
+  }
 
+  render() {
+    return (
+      <div>
+        <div className='container'>
+        {React.cloneElement(this.props.children, {setLocation: this.setLocation.bind(this)})}
+        </div>
+      </div>
+    );
+  }
+};
+
+export default App;
