+++ client/routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import App from './app/app.jsx';
+import App from './app/app.jsx';
 import Decision from './app/decision.jsx';
 import Location from './app/location.jsx';
 import mapBar from './app/mapBar.jsx';
@@ -10,31 +10,6 @@ import signup from './app/signup.jsx';
 import Index from './app/index.jsx';
 import { Router, Route, IndexRoute, hashHistory } from 'react-router';
 
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      location: 'San Francisco'
-    };
-  }
-
-  setLocation(loc) {
-    this.setState({
-      location: loc
-    });
-  }
-
-  render() {
-    return (
-      <div>
-        <div className='container'>
-        {React.cloneElement(this.props.children, {setLocation: this.setLocation.bind(this)})}
-        </div>
-      </div>
-    );
-  }
-};
 
 
 ReactDOM.render (
